feat(Platform,Player): Add cross-platform features
- Add Linux build and runtime support
- Fix SDL window close and rendering
issues on Linux
- Implement Windows taskbar multimedia control
- Add system
volume/mixer control to Windows playback options
- Change size_t to uint64_t
for type consistency

+++ QtSDLFFmpegVideoPlayer/Audio/AudioAdapter/AudioAdapter.h
@@ -90,7 +90,7 @@ public:
         AFFloat32,
         AFFloat64
     };
-    constexpr static size_t AudioFormatCount = sizeof(AudioFormatList) / sizeof(AudioFormatList[0]);
+    constexpr static uint64_t AudioFormatCount = sizeof(AudioFormatList) / sizeof(AudioFormatList[0]);
 
     using AudioFormats = MultiEnumTypeDefine<AudioFormat>;
