+++ QtSDLFFmpegVideoPlayer/Utils/ThreadUtils.h
@@ -26,7 +26,7 @@
 ////        std::this_thread::yield();  // 让出CPU时间片，降低资源占用
 ////}
 //
-//inline void SleepForMs(size_t duration)
+//inline void SleepForMs(uint64_t duration)
 //{
 //    auto to = std::chrono::steady_clock::now() + std::chrono::milliseconds(duration);
 //    while (std::chrono::steady_clock::now() < to)
@@ -34,7 +34,7 @@
 //}
 //
 //
-//inline void SleepForUs(size_t duration)
+//inline void SleepForUs(uint64_t duration)
 //{
 //    auto tp = std::chrono::high_resolution_clock::now() + std::chrono::microseconds(duration);
 //    while (std::chrono::high_resolution_clock::now() < tp)
@@ -48,23 +48,23 @@ inline void SleepFor(const std::chrono::duration<_Rep, _Period>& duration)
     std::this_thread::sleep_for(duration);
 }
 
-inline void SleepForS(size_t duration)
+inline void SleepForS(uint64_t duration)
 {
     SleepFor(std::chrono::seconds(duration));
 }
 
-inline void SleepForMs(size_t duration)
+inline void SleepForMs(uint64_t duration)
 {
     SleepFor(std::chrono::milliseconds(duration));
 }
 
 
-inline void SleepForUs(size_t duration)
+inline void SleepForUs(uint64_t duration)
 {
     SleepFor(std::chrono::microseconds(duration));
 }
 
-inline void SleepForNs(size_t duration)
+inline void SleepForNs(uint64_t duration)
 {
     SleepFor(std::chrono::nanoseconds(duration));
 }
